Add javadocs to project to fix gradle build with checkstyle
