+++ src/main/java/duke/exceptions/GeneralException.java
@@ -1,11 +1,16 @@
 package duke.exceptions;
 
+/**
+ * Represents a general exception that is thrown when an unknown error occurs during program execution
+ * in the DailyTasks application.
+ * This exception can be used as a fallback for unexpected errors.
+ */
 public class GeneralException extends Exception {
 
     /**
-     * Creates a duke.exceptions.GeneralException when an unknown exception happens during program execution.
+     * Constructs a GeneralException with a specified error message.
      *
-     * @param e description of the exception
+     * @param e The error message describing the exception.
      */
     public GeneralException(String e) {
         super(e);
