+++ src/main/java/duke/exceptions/InvalidEventException.java
@@ -1,11 +1,15 @@
 package duke.exceptions;
 
+/**
+ * Represents an exception that is thrown when a user enters an invalid event in the DailyTasks application.
+ * This exception typically occurs when the `from` or `to` time for the event is incorrectly formatted or invalid.
+ */
 public class InvalidEventException extends Exception {
+
     /**
-     * Creates a duke.exceptions.InvalidEventException when the user enters an event with an invalid
-     * `from` or `to` time.
+     * Constructs an InvalidEventException with a specified error message.
      *
-     * @param e description of the exception
+     * @param e The error message describing the exception.
      */
     public InvalidEventException(String e) {
         super(e);
