+++ src/main/java/duke/exceptions/UnknownMessageException.java
@@ -1,16 +1,23 @@
 package duke.exceptions;
 
+/**
+ * Represents an exception that is thrown when the user enters a message that is not recognized as a valid task in
+ * the DailyTasks application.
+ * This exception is used to signal that the input does not correspond to any known command or task.
+ */
 public class UnknownMessageException extends Exception {
+
     /**
-     * Creates a duke.exceptions.UnknownMessageException when the user enters a message that is not a task.
-     *
+     * Constructs an UnknownMessageException with no error message.
+     * This is used when the exception needs to be thrown without a specific error message.
      */
     public UnknownMessageException() {}
 
     /**
-     * Creates a duke.exceptions.UnknownMessageException when the user enters a message that is not a task.
+     * Constructs an UnknownMessageException with a specified error message.
+     * This constructor is used when additional details about the exception are needed.
      *
-     * @param e description of the exception
+     * @param e The error message describing the exception.
      */
     public UnknownMessageException(String e) {
         super(e);
