+++ src/main/java/duke/storage/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.storage;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -16,12 +16,26 @@ import duke.tasks.Event;
 import duke.tasks.Task;
 import duke.tasks.ToDos;
 
+/**
+ * The Storage class handles saving and loading tasks from a file to persist the state of the task list.
+ * It reads from and writes to a specified file to store the tasks across sessions.
+ */
 public class Storage {
+
+    /** The directory where the state file is stored. */
     private static final String STATE_FILE_DIRECTORY = "./data";
+
+    /** The name of the state file where tasks are saved. */
     private static final String STATE_FILE = "save.txt";
 
+    /**
+     * Converts a line of text from the state file into a Task object.
+     *
+     * @param state The line of text representing a saved task.
+     * @return The corresponding Task object, or null if there is an error converting the task.
+     */
     private static Task convertStateToTask(String state) {
-        // format of state file below
+        // Format of state file:
         // T | 1 | read book
         // D | 0 | return book | June 6th
         // E | 0 | project meeting | Aug 6th 2-4pm
@@ -60,6 +74,12 @@ public class Storage {
         return task;
     }
 
+    /**
+     * Saves the current list of tasks to the state file.
+     *
+     * @param tasks The list of tasks to be saved.
+     * @throws IOException If an I/O error occurs while writing to the file.
+     */
     public static void saveTasksListToStateFile(List<Task> tasks) throws IOException {
         Path dirPath = Paths.get(Storage.STATE_FILE_DIRECTORY);
         Path filePath = dirPath.resolve(Storage.STATE_FILE);
@@ -79,6 +99,12 @@ public class Storage {
         writer.close();
     }
 
+    /**
+     * Converts a Task object into a string that can be saved in the state file.
+     *
+     * @param task The task to be converted.
+     * @return The string representation of the task for storage.
+     */
     private static String convertTaskToState(Task task) {
         StringBuilder str = new StringBuilder();
         str.append(task.getStatusIcon().equals("X") ? "| 1 " : "| 0 ");
@@ -103,15 +129,22 @@ public class Storage {
         return str.toString();
     }
 
+    /**
+     * Loads the state file and converts its content to a list of Task objects.
+     * Creates the state file and directory if they do not exist.
+     *
+     * @return The list of tasks read from the state file.
+     * @throws IOException If an I/O error occurs while reading the file.
+     */
     public List<Task> loadStateFileToTasksList() throws IOException {
         Path dirPath = Paths.get(Storage.STATE_FILE_DIRECTORY);
         Path filePath = dirPath.resolve(Storage.STATE_FILE);
 
-        if (Files.notExists(dirPath)) { // create the SAVE_DIRECTORY folder if it doesn't exist
+        if (Files.notExists(dirPath)) { // Create the directory if it doesn't exist
             Files.createDirectories(dirPath);
         }
 
-        if (Files.notExists(filePath)) { // create an empty file initially
+        if (Files.notExists(filePath)) { // Create an empty file initially if it doesn't exist
             Files.createFile(filePath);
         }
 
@@ -123,7 +156,7 @@ public class Storage {
         }
         reader.close();
 
-        // convert the state file `lines` to an arraylist of tasks
+        // Convert the state file lines to a list of tasks
         List<Task> tasks = new ArrayList<>();
         for (String state : lines) {
             tasks.add(Storage.convertStateToTask(state));
